Read in the simulation, too
